requirements.txt
requests>=2.31,<3.0
feedparser>=6.0,<7.0
beautifulsoup4>=4.12,<5.0
text-dedup>=0.3,<1.0
sentence-transformers>=3.0,<4.0
pydantic>=2.0,<3.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
pyTelegramBotAPI>=4.15.0
